requirements.txt
python-dotenv
requests
groq
sentence-transformers
chromadb
SPARQLWrapper
